Update CI/CD workflows and build configurations
This revision modifies the GitHub CI/CD workflows (release.yml, main.yml, and integration.yml) as well as the build configurations. Notably, a more comprehensive syntax is employed in the workflows for caching and task execution, an adjusted cache key strategy is implemented, and the inclusion of new package references in the build configuration is done.