Update package versions and add OptionParserException
Updated the versions for numerous packages throughout several project files ensuring the project is running on the most recent dependencies. Also added a new class, OptionParserException, which throws an exception when reading option properties fails. This commits further includes some minor changes in the implementation of argument checks in OptionParser.